feature first, tests in separe file and refactor on comparations
